pipeline-manager: split connector-specific DB calls into a separate module
